Convert DOS Lineendings to UNIX
In case you use windows and your git tool converts the files to windows lineendings, this should convert the entrypoint.sh back to UNIX lineendings.
I wanted to test this, but gave up because I have not enough patience for this horrible OS. Windows pisses me of, makes me mad and is starting to ruin my day, so fuck it.